feat: improve navigation responsiveness in the protocols page
