+++ app/dashboard/participants/_components/ExportParticipants/ImportExportSection.tsx
@@ -1,15 +1,15 @@
-import Paragraph from '~/components/ui/typography/Paragraph';
-import SettingsSection from '~/components/layout/SettingsSection';
-import ImportCSVModal from '../ImportCSVModal';
-import ExportParticipants, {
-  ExportParticipantsFallback,
-} from './ExportParticipants';
 import ResponsiveContainer from '~/components/ResponsiveContainer';
-import { Suspense } from 'react';
-import { getParticipants } from '~/queries/participants';
+import SettingsSection from '~/components/layout/SettingsSection';
+import Paragraph from '~/components/ui/typography/Paragraph';
+import type { GetParticipantsReturnType } from '~/queries/participants';
+import ImportCSVModal from '../ImportCSVModal';
+import ExportParticipants from './ExportParticipants';
 
-export default function ImportExportSection() {
-  const participantsPromise = getParticipants();
+export default function ImportExportSection({
+  participantsPromise,
+}: {
+  participantsPromise: GetParticipantsReturnType;
+}) {
   return (
     <ResponsiveContainer>
       <SettingsSection
@@ -17,9 +17,7 @@ export default function ImportExportSection() {
         controlArea={
           <div className="flex w-72 flex-col items-center justify-end gap-4">
             <ImportCSVModal />
-            <Suspense fallback={<ExportParticipantsFallback />}>
-              <ExportParticipants participantsPromise={participantsPromise} />
-            </Suspense>
+            <ExportParticipants participantsPromise={participantsPromise} />
           </div>
         }
       >
