+++ app/dashboard/participants/page.tsx
@@ -1,26 +1,23 @@
-import ParticipantsTable from '~/app/dashboard/_components/ParticipantsTable/ParticipantsTable';
 import ResponsiveContainer from '~/components/ResponsiveContainer';
 import Section from '~/components/layout/Section';
-import PageHeader from '~/components/ui/typography/PageHeader';
-import { requireAppNotExpired } from '~/queries/appSettings';
-import { requirePageAuth } from '~/utils/auth';
+import { getParticipants } from '~/queries/participants';
+import { getProtocols } from '~/queries/protocols';
+import { ParticipantsTableClient } from '../_components/ParticipantsTable/ParticipantsTableClient';
 import ImportExportSection from './_components/ExportParticipants/ImportExportSection';
 
-export default async function ParticipantPage() {
-  await requireAppNotExpired();
-  await requirePageAuth();
+export default function ParticipantPage() {
+  const participantsPromise = getParticipants();
+  const protocolsPromise = getProtocols();
+
   return (
     <>
-      <ResponsiveContainer>
-        <PageHeader
-          headerText="Participants"
-          subHeaderText="View and manage your participants."
-        />
-      </ResponsiveContainer>
-      <ImportExportSection />
+      <ImportExportSection participantsPromise={participantsPromise} />
       <ResponsiveContainer maxWidth="6xl">
         <Section>
-          <ParticipantsTable />
+          <ParticipantsTableClient
+            participantsPromise={participantsPromise}
+            protocolsPromise={protocolsPromise}
+          />
         </Section>
       </ResponsiveContainer>
     </>
